+++ backend/canisters/user_index/impl/src/queries/users.rs
@@ -16,6 +16,7 @@ fn users_impl(args: Args, state: &RuntimeState) -> Response {
 
     let mut user_ids = HashSet::new();
     let mut users = Vec::new();
+    let mut deleted = Vec::new();
     let mut current_user: Option<CurrentUserSummary> = None;
 
     if let Some(u) = state.data.users.get_by_principal(&caller) {
@@ -50,24 +51,27 @@ fn users_impl(args: Args, state: &RuntimeState) -> Response {
     let now_month = MonthKey::from_timestamp(now);
     for group in args.user_groups {
         let updated_since = group.updated_since;
-        users.extend(
-            group
-                .users
-                .into_iter()
-                .filter_map(|u| state.data.users.get_by_user_id(&u))
-                .filter(move |u| {
-                    (u.date_updated > updated_since
-                        || u.chit_updated > updated_since
-                        || (now > u.streak_ends && u.streak_ends > updated_since))
-                        && u.principal != caller
-                })
-                .filter(|u| user_ids.insert(u.user_id))
-                .map(|u| UserSummaryV2 {
-                    user_id: u.user_id,
-                    stable: (u.date_updated > updated_since).then(|| u.to_summary_stable(now)),
-                    volatile: Some(u.to_summary_volatile(now, now_month)),
-                }),
-        );
+
+        for user_id in group.users {
+            if !user_ids.insert(user_id) {
+                continue;
+            }
+            if let Some(user) = state.data.users.get_by_user_id(&user_id).filter(|u| {
+                (u.date_updated > updated_since
+                    || u.chit_updated > updated_since
+                    || (now > u.streak_ends && u.streak_ends > updated_since))
+                    && u.principal != caller
+            }) {
+                users.push(UserSummaryV2 {
+                    user_id,
+                    stable: (user.date_updated > updated_since).then(|| user.to_summary_stable(now)),
+                    volatile: Some(user.to_summary_volatile(now, now_month)),
+                });
+                // TODO maybe convert `deleted_users` to a HashMap?
+            } else if state.data.deleted_users.iter().any(|u| u.user_id == user_id) {
+                deleted.push(user_id)
+            }
+        }
     }
 
     if let Some(ts) = args.users_suspended_since {
@@ -87,6 +91,7 @@ fn users_impl(args: Args, state: &RuntimeState) -> Response {
     Success(Result {
         users,
         current_user,
+        deleted,
         timestamp: now,
     })
 }